mavlink: fix on target unit tests
This fixes the unit tests on target, e.g. px4_fmu-v2_test.

The problem was that no MAVLink instances are instantiated and hence
mavlink_status_t for COMM_0 is null. When the mavlink message is packed
it accesses the status->signing which is then garbage and leads to a
segfault.

The fix involves using the MAVLink headers without convenience functions
defined for the unit tests and therefore not trying to use data from the
mavlink module.